Login dialog action: Fix login error.
If a user has an active session and tries to login in a diffeent
browser then user is shown pop-up to cancel or login and log-off other sessions.

If user chooses to log-off other session and that session is terminated
in between. This request causes server error as it is assumed that there
is always another active session when user chooses to log-off other
sessions and login.